requirements.txt
pandas
streamlit
scikit-learn
xgboost
catboost
plotly
Pillow
streamlit-lottie